allow pod related metrics to be filtered by its phase
Currently, the graph considers pods that have been terminated.
This leads to confusing reporting, and sometimes we saw that the Global CPU Usage
could go over 100% and goes even as high as 200%.
For our case, it was due to a lot of short running cronjob-spawned pods
that are terminated one after another.

The proposes a new 'Pod Phase' filter which defaults to 'Running' & 'Pending' pods,
and users can choose to see All pods depending on their needs.