feat(k8s-views-nodes): Support FQDN nodename
It is fairly common for the server hostname to be a FQDN (e.g.
`node-01.example.com`), while the node is known to Kubernetes by it's
short name (e.g. `node-01`).

By relaxing the regex that matches the `nodename` label to determine the
`instance` name, we can cover that case without the need to relabel
metrics.

Signed-off-by: Benoît Knecht <bknecht@protonmail.ch>